Fix the dialog positioning being wrong on iOS
So, we would make the dialog visible, and then we calculate its
positioning based on its height and the scroll position.
Then we would get the scroll position and page height and set a bunch of
properties to prevent page scrolling.

iOS automatically scrolls the page when an input gains focus.
When the dialog is made visible, the textarea gains focus.

This caused iOS to scroll the page in between us showing the dialog and
starting the positioning, giving us weirdly huge scroll offsets that
resulted in the dialog being mostly off-screen.

Now, we grab the scroll offset once, before making the dialog visible
and use that cached value for both positioning and scroll-blocking.