Terrible solution for autofocus inputs on iOS
Let's invent our own attribute rather than using the one in the spec,
because Safari actually obeys the spec and we can't prevent that.

We wrap the focus in a timeout, because that causes Safari to ignore it
(due to not being user interaction).

Wait, now we have to hack around Chrome, which supports <dialog>
natively and needs the autofocus attribute to work properly. So in that
case we'll let it autofocus the first button, and immediately call our
own setFocus function to find out dialog-autofocus element.
We need to do that without a timeout, otherwise we get some confusing
focus styling on the close button temporarily.