Use a sentinel node to ensure dialogs close
For positioning and accessibility purposes, when we open the dialog on a
browser that doesn't support <dialog> natively, we end up moving it up
the tree to be a child of <body>.

Unfortunately, that causes cases where the original parent node of the
dialog is removed from the DOM without causing the dialog to close.

Our solution is to leave a placeholder sentinel node in the DOM at the
original location of the <dialog> element, and check whenever the DOM is
modified to see if the body still contains that sentinel element. When
it is removed, we need to close the dialog associated with the sentinel.

This was most noticeable in Safari on iOS where you could end up with
full-screen modals that didn't close, even as the page navigated away
underneath them.