Fixes for weird WKWebView bug with safe areas
When opening a modal, if the body changes to position: fixed, then the
safe-areas start affecting content regardless of the viewport-fit meta
directive. In theory, it's highly unlikely that we need position: fixed
over something like position: absolute, since it's exceedingly rare for
there to be unexpected relative positioning on the html element.